+++ models/MoviesDao.js
@@ -1,13 +1,7 @@
 const { PrismaClient } = require("@prisma/client");
 const prisma = new PrismaClient();
 
-const getGenres = async () => {
-  return await prisma.$queryRaw`
-		SELECT id, genre_name FROM movies_genre; 
-	`;
-};
-
-const getMoviesByGenre = async genreName => {
+const getMoviesByGenre = async genreId => {
   return await prisma.$queryRaw`
     SELECT
     movies.id AS id,
@@ -15,10 +9,10 @@ const getMoviesByGenre = async genreName => {
     movies.poster_url AS poster_url
     FROM movies
     JOIN movies_genre ON movies.genre_id = movies_genre.id 
-    WHERE movies_genre.genre_name = ${genreName}
+    WHERE movies_genre.id = ${genreId}
   `;
 };
+
 module.exports = {
-  getGenres,
   getMoviesByGenre,
 };
