+++ routes/index.js
@@ -2,8 +2,8 @@ const express = require("express");
 const router = express.Router();
 
 const testRouter = require("./testRoute");
-const movieRouter = require("./movieRoute");
+const moviesRouter = require("./MoviesRoute");
 
 router.use("/test", testRouter);
-router.use("/movie", movieRouter);
+router.use("/movies", moviesRouter);
 module.exports = router;
