--- routes/movieRoute.js
@@ -1,9 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-const SimilarMovieController = require("../controllers/SimilarMovieController");
-
-
-router.get("/",SimilarMovieController.getMoviesByGenre);
-router.get("/genre",SimilarMovieController.getGenres);
-module.exports = router;
